refactor: move a dummy data to a separate file
