+++ src/helpers/util.ts
@@ -11,3 +11,21 @@ export const calculateServiceDuration = (duration: number) => {
 
   return serviceDuration;
 };
+
+export const sortByDate = (dates: string[]) => dates.sort((a, b) => new Date(a).getTime() - new Date(b).getTime());
+
+export const splitToDays = (dates: string[]) => {
+  const datesToReturn = [[dates[0]]];
+  let k = 0;
+
+  for (let i = 1; i < dates.length; i++) {
+    if (datesToReturn[k][0].split('T')[0] === dates[i].split('T')[0]) {
+      datesToReturn[k].push(dates[i]);
+    } else {
+      k++;
+      datesToReturn[k] = [dates[i]];
+    }
+  }
+
+  return datesToReturn;
+};
