+++ src/views/Salon/Components/Reservation/reservation.data.ts
@@ -1,0 +1,26 @@
+// it's a temporary file and will be removed after connecting the reservation view to backend
+export const freeHours = {
+  freeHours: [
+    {
+      workedId: '60ae96dad2396a054c3ab53e',
+      startHours: [
+        '2021-10-04T13:15:00.000Z',
+        '2021-10-03T12:45:00.000Z',
+        '2021-10-03T12:15:00.000Z',
+        '2021-10-03T12:30:00.000Z',
+      ],
+    },
+    {
+      workedId: '60ae96dad2396a054c3ab53f',
+      startHours: ['2021-10-03T12:45:00.000Z', '2021-10-03T12:30:00.000Z', '2021-10-03T13:00:00.000Z'],
+    },
+    {
+      workedId: '60ae96dad2396a054c3ab53g',
+      startHours: ['2021-10-04T13:00:00.000Z', '2021-10-03T12:15:00.000Z', '2021-10-03T12:30:00.000Z'],
+    },
+    {
+      workedId: '60ae96dad2396a054c3ab53g',
+      startHours: ['2021-10-03T12:15:00.000Z', '2021-10-03T12:45:00.000Z', '2021-10-04T13:00:00.000Z'],
+    },
+  ],
+};
